+++ src/BpsPatch.Core/GlobalUsings.cs
@@ -13,3 +13,5 @@ global using System.IO.Hashing;
 global using System.Numerics;
 global using System.Runtime.CompilerServices;
 global using System.Text;
+global using System.Threading;
+global using System.Threading.Tasks;